TheaterLink/requirements.txt
Flask==3.0.2
psycopg2==2.9.9
python-dotenv==1.0.1
rsa==4.7.2
cryptography==36.0.0
